Add in conf.d/downloads file to download latest
Also tweak conf.d/main to use the downloaded file.

Previously we used Moosh to download the latest Moodle version but that
fails to download the latest 3.9.0 version for some reason?

I was thinking of using our gh_releases script to pul lthe latest
release, but due to the range of URLs to check (Moodle release patch
releases every ~week) I decdied to just hardcode it for now...

Not sure if we'll continue this path, but we'll leave it be for now...